+++ src/deflate.rs
@@ -15,7 +15,7 @@
 //! use gzp::{deflate::Zlib, par::compress::{ParCompress, ParCompressBuilder}, ZWriter};
 //!
 //! let mut writer = vec![];
-//! let mut parz: ParCompress<Zlib> = ParCompressBuilder::new().from_writer(writer);
+//! let mut parz: ParCompress<Zlib,_> = ParCompressBuilder::new().from_writer(writer);
 //! parz.write_all(b"This is a first test line\n").unwrap();
 //! parz.write_all(b"This is a second test line\n").unwrap();
 //! parz.finish().unwrap();
@@ -153,10 +153,9 @@ where
     }
 }
 
-impl<W: Write> ZWriter for SyncZ<GzEncoder<W>> {
-    fn finish(&mut self) -> Result<(), GzpError> {
-        self.inner.take().unwrap().finish()?;
-        Ok(())
+impl<W: Write> ZWriter<W> for SyncZ<GzEncoder<W>> {
+    fn finish(&mut self) -> Result<W, GzpError> {
+        Ok(self.inner.take().unwrap().finish()?)
     }
 }
 
@@ -265,10 +264,9 @@ where
 }
 
 #[cfg(feature = "any_zlib")]
-impl<W: Write> ZWriter for SyncZ<ZlibEncoder<W>> {
-    fn finish(&mut self) -> Result<(), GzpError> {
-        self.inner.take().unwrap().finish()?;
-        Ok(())
+impl<W: Write> ZWriter<W> for SyncZ<ZlibEncoder<W>> {
+    fn finish(&mut self) -> Result<W, GzpError> {
+        Ok(self.inner.take().unwrap().finish()?)
     }
 }
 
@@ -344,10 +342,9 @@ where
     }
 }
 
-impl<W: Write> ZWriter for SyncZ<DeflateEncoder<W>> {
-    fn finish(&mut self) -> Result<(), GzpError> {
-        self.inner.take().unwrap().finish()?;
-        Ok(())
+impl<W: Write> ZWriter<W> for SyncZ<DeflateEncoder<W>> {
+    fn finish(&mut self) -> Result<W, GzpError> {
+        Ok(self.inner.take().unwrap().finish()?)
     }
 }
 
@@ -494,10 +491,9 @@ where
     }
 }
 
-impl<W: Write> ZWriter for SyncZ<MgzipSyncWriter<W>> {
-    fn finish(&mut self) -> Result<(), GzpError> {
-        self.inner.take().unwrap().flush()?;
-        Ok(())
+impl<W: Write> ZWriter<W> for SyncZ<MgzipSyncWriter<W>> {
+    fn finish(&mut self) -> Result<W, GzpError> {
+        Ok(self.inner.take().unwrap().finish()?)
     }
 }
 
@@ -650,10 +646,9 @@ where
     }
 }
 
-impl<W: Write> ZWriter for SyncZ<BgzfSyncWriter<W>> {
-    fn finish(&mut self) -> Result<(), GzpError> {
-        self.inner.take().unwrap().flush()?;
-        Ok(())
+impl<W: Write> ZWriter<W> for SyncZ<BgzfSyncWriter<W>> {
+    fn finish(&mut self) -> Result<W, GzpError> {
+        Ok(self.inner.take().unwrap().finish()?)
     }
 }
 
@@ -695,7 +690,7 @@ mod test {
         ";
 
         // Compress input to output
-        let mut par_gz: ParCompress<Mgzip> = ParCompressBuilder::new().from_writer(out_writer);
+        let mut par_gz: ParCompress<Mgzip, _> = ParCompressBuilder::new().from_writer(out_writer);
         par_gz.write_all(input).unwrap();
         par_gz.finish().unwrap();
 
@@ -728,7 +723,7 @@ mod test {
         ";
 
         // Compress input to output
-        let mut par_gz: ParCompress<Gzip> = ParCompressBuilder::new().from_writer(out_writer);
+        let mut par_gz: ParCompress<Gzip, _> = ParCompressBuilder::new().from_writer(out_writer);
         par_gz.write_all(input).unwrap();
         par_gz.finish().unwrap();
 
@@ -761,7 +756,7 @@ mod test {
         ";
 
         // Compress input to output
-        let mut par_gz: ParCompress<Gzip> = ParCompressBuilder::new().from_writer(out_writer);
+        let mut par_gz: ParCompress<Gzip, _> = ParCompressBuilder::new().from_writer(out_writer);
         par_gz.write_all(input).unwrap();
         drop(par_gz);
 
@@ -861,9 +856,46 @@ mod test {
         ";
 
         // Compress input to output
-        let mut par_gz: ParCompress<Zlib> = ParCompressBuilder::new().from_writer(out_writer);
+        let mut par_gz: ParCompress<Zlib, _> = ParCompressBuilder::new().from_writer(out_writer);
         par_gz.write_all(input).unwrap();
         par_gz.finish().unwrap();
+
+        // Read output back in
+        let mut reader = BufReader::new(File::open(output_file).unwrap());
+        let mut result = vec![];
+        reader.read_to_end(&mut result).unwrap();
+
+        // Decompress it
+        let mut gz = ZlibDecoder::new(&result[..]);
+        let mut bytes = vec![];
+        gz.read_to_end(&mut bytes).unwrap();
+
+        // Assert decompressed output is equal to input
+        assert_eq!(input.to_vec(), bytes);
+    }
+
+    #[test]
+    #[cfg(feature = "any_zlib")]
+    fn test_scoped_simple_zlib() {
+        let dir = tempdir().unwrap();
+
+        // Create output file
+        let output_file = dir.path().join("output.txt");
+        let mut out_writer = BufWriter::new(File::create(&output_file).unwrap());
+
+        // Define input bytes
+        let input = b"\
+        This is a longer test than normal to come up with a bunch of text.\n\
+        We'll read just a few lines at a time.\n\
+        ";
+
+        std::thread::scope(|s| {
+            // Compress input to output
+            let mut par_gz: ParCompress<Zlib, _> =
+                ParCompressBuilder::new().from_borrowed_writer(&mut out_writer, s);
+            par_gz.write_all(input).unwrap();
+            par_gz.finish().unwrap();
+        });
 
         // Read output back in
         let mut reader = BufReader::new(File::open(output_file).unwrap());
@@ -938,7 +970,7 @@ mod test {
         ];
 
         // Compress input to output
-        let mut par_gz: ParCompress<Gzip> = ParCompressBuilder::new()
+        let mut par_gz: ParCompress<Gzip, _> = ParCompressBuilder::new()
             .buffer_size(DICT_SIZE)
             .unwrap()
             .from_writer(out_writer);
@@ -974,7 +1006,7 @@ mod test {
         ";
 
         // Compress input to output
-        let mut par_gz: ParCompress<Mgzip> = ParCompressBuilder::new().from_writer(out_writer);
+        let mut par_gz: ParCompress<Mgzip, _> = ParCompressBuilder::new().from_writer(out_writer);
         par_gz.write_all(input).unwrap();
         par_gz.finish().unwrap();
 
@@ -1003,7 +1035,7 @@ mod test {
         ";
 
         // Compress input to output
-        let mut par_gz: ParCompress<Bgzf> = ParCompressBuilder::new().from_writer(out_writer);
+        let mut par_gz: ParCompress<Bgzf, _> = ParCompressBuilder::new().from_writer(out_writer);
         par_gz.write_all(input).unwrap();
         par_gz.finish().unwrap();
 
@@ -1040,7 +1072,7 @@ mod test {
 
 
             // Compress input to output
-            let mut par_gz: Box<dyn ZWriter> = if num_threads > 0 {
+            let mut par_gz: Box<dyn ZWriter<_>> = if num_threads > 0 {
                 Box::new(ParCompressBuilder::<Gzip>::new()
                     .buffer_size(buf_size).unwrap()
                     .num_threads(num_threads).unwrap()
@@ -1086,7 +1118,7 @@ mod test {
 
 
             // Compress input to output
-            let mut par_gz: Box<dyn ZWriter> = if num_threads > 0 {
+            let mut par_gz: Box<dyn ZWriter<_>> = if num_threads > 0 {
                 Box::new(ParCompressBuilder::<Mgzip>::new()
                     .buffer_size(buf_size).unwrap()
                     .num_threads(num_threads).unwrap()
@@ -1183,7 +1215,7 @@ mod test {
             }
 
             // Compress input to output
-            let mut par_gz: Box<dyn ZWriter> = if num_threads > 0 {
+            let mut par_gz: Box<dyn ZWriter<_>> = if num_threads > 0 {
                 Box::new(ParCompressBuilder::<Bgzf>::new()
                     .buffer_size(buf_size).unwrap()
                     .num_threads(num_threads).unwrap()
@@ -1316,7 +1348,7 @@ mod test {
 
 
             // Compress input to output
-            let mut par_gz: Box<dyn ZWriter> = if num_threads > 0 {
+            let mut par_gz: Box<dyn ZWriter<_>> = if num_threads > 0 {
                 Box::new(ParCompressBuilder::<Zlib>::new()
                     .buffer_size(buf_size).unwrap()
                     .num_threads(num_threads).unwrap()