+++ src/mgzip.rs
@@ -92,7 +92,7 @@ where
     #[cfg(not(feature = "libdeflate"))]
     compressor: Compress,
     /// The inner writer
-    writer: W,
+    writer: Option<W>,
 }
 
 impl<W> MgzipSyncWriter<W>
@@ -116,8 +116,15 @@ where
             blocksize,
             compression_level,
             compressor,
-            writer,
+            writer: Some(writer),
         }
+    }
+
+    pub(crate) fn finish(mut self) -> io::Result<W> {
+        self.flush()?;
+        self.writer
+            .take()
+            .ok_or_else(|| io::Error::new(io::ErrorKind::Other, "Writer already taken"))
     }
 }
 
@@ -288,20 +295,23 @@ where
             let b = self.buffer.split_to(self.blocksize).freeze();
             let compressed = compress(&b[..], &mut self.compressor, self.compression_level)
                 .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
-            self.writer.write_all(&compressed)?;
+            self.writer.as_mut().unwrap().write_all(&compressed)?;
         }
         Ok(buf.len())
     }
 
     /// Flush this output stream, ensuring all intermediately buffered contents are sent.
     fn flush(&mut self) -> std::io::Result<()> {
-        let b = self.buffer.split_to(self.buffer.len()).freeze();
-        if !b.is_empty() {
-            let compressed = compress(&b[..], &mut self.compressor, self.compression_level)
-                .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
-            self.writer.write_all(&compressed)?;
+        if let Some(writer) = self.writer.as_mut() {
+            let b = self.buffer.split_to(self.buffer.len()).freeze();
+            if !b.is_empty() {
+                let compressed = compress(&b[..], &mut self.compressor, self.compression_level)
+                    .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+                writer.write_all(&compressed)?;
+            }
+            writer.flush()?;
         }
-        self.writer.flush()
+        Ok(())
     }
 }
 
@@ -310,7 +320,7 @@ where
     W: Write,
 {
     fn drop(&mut self) {
-        self.flush().unwrap();
+        let _ = self.flush();
     }
 }